requirements.txt
streamlit==1.45.0
google-generativeai==0.8.5
pandas==2.2.3
python-dotenv==1.1.0
SpeechRecognition==3.14.3
